+++ code/editor/QuickCompile.h
@@ -69,7 +69,7 @@ public:
 	// ClassWizard generated virtual function overrides
 	//{{AFX_VIRTUAL(CQuickCompile)
 	public:
-	virtual int DoModal();
+	virtual intptr_t DoModal();
 	protected:
 	virtual void DoDataExchange(CDataExchange* pDX);    // DDX/DDV support
 	//}}AFX_VIRTUAL
